Modify actor to accept a policy function
Modify the actor object to accept a policy function on initialization
instead of a policy network.
The policy function returns a probability distribution over the action
space and the actor samples from that distribution.
This change facilitates the implementation of other agents, such as DQN
where there is no policy network.